plumbing: transport, refactor transport Endpoint to use url.URL
There's no need to duplicate the logic already present in url.URL and
the standard library to parse and handle URLs. This refactor simplifies
the codebase by removing redundant fields and methods, and leverages the
existing functionality of url.URL.

This also remove unused code in transport/http ModifyEndpointIfRedirect
and its related tests.