plumbing: improve revfile encode/decode
related to #1731

* Make encode and decode a function and unexport
encoder/decoder to hide state so that
there is less opportunity to misuse concurrently,
and drop mutex and recover as a result.
* Handle non EOF errors in decode trailing check.
* Add context in object decode for EOF.
* Drop unused stateFn in encoder/decoder.
* Accept io.Reader since it's more idiomatic.
For decode, internally, we create buffered reader
if already not buffered but not doing it for
writing since flushing is required and it's
generally controlled by caller (file, buffer, etc).

Keeping explicit check for io.EOF because of
convention https://github.com/golang/go/issues/39155

Signed-off-by: ferhat elmas <elmas.ferhat@gmail.com>