utils: ioutil, Avoid allocation in context I/O
To protect the buffer that was passed to the context reader/wrtier, it must be
copied into a separate buffer. The previous implementation allocated a new
buffer of the same size for each call. This results in poor performance due to
excessive memory allocation. Instead, replace the per-call allocation with a
buffer obtained from a sync.Pool.