plumbing: fix opening relative paths
go-billy checks that in a chroot filesystem the paths do not start by
`../`. This causes opening or cloning relative paths to fail. This check
is good to find repository links that point to files outside the
repository but for opening path is not that useful as the default
filesystem is based at `/`.

The change to `parseFile` does the same as v5. The test and the embedded
function to generate the relative path is quite naive but checks that
the change fixes the issue.

Fixes #1723

Signed-off-by: Javi Fontan <jfontan@gmail.com>