git: status, Use head tree on Status to determine unmodified file
To determine whether a file is unmodified, previous implementaion preloaded
all the entries of the index to the status map. Instead, we can use head tree
directly from Status. This way we can allign with upstream git while maintaining
small amount of elements in status map.

How it works:
1. The status map only consists of the statuses of files that are changed.
2. If a file doesn't exist in status map, it's either not modified or invisible to git.
3. If the file from step 2 exists in head tree, it's not modified(=unmodified).
4. If not, it's invisible to git(=untracked).

Since head tree is now included in Status, it had to be changed to a struct.
This resulted in changes of the API. Changes include:
1. Addition of Iter() and Len() as the map is now unexported.
2. Removal of pointer on FileStatus as it's unnecessary.

Unlike previous implementation where determining unmodified is opt-in method,
now it is the one and only behavior for Status(). This lead to test failures as
some of them depended on how previous default behavior of Status(). I fixed
those tests as well by replacing 'Untracked' with 'Unmodified'.